Moved DiagnoticException in a dedicate cs file
